move MLIPPredictUnit
