+++ src/fairchem/core/calculate/pretrained_mlip.py
@@ -16,10 +16,10 @@ from huggingface_hub import hf_hub_download
 
 from fairchem.core import calculate
 from fairchem.core._config import CACHE_DIR
-from fairchem.core.units.mlip_unit import load_predict_unit
+from fairchem.core.units.mlip_unit import MLIPPredictUnit, load_predict_unit
 
 if TYPE_CHECKING:
-    from fairchem.core.units.mlip_unit import InferenceSettings, MLIPPredictUnit
+    from fairchem.core.units.mlip_unit import InferenceSettings
 
 
 @dataclass
