+++ tests/core/units/mlip_unit/test_inference_checkpoint.py
@@ -23,7 +23,8 @@ from fairchem.core import FAIRChemCalculator
 from fairchem.core.datasets.ase_datasets import AseDBDataset
 from fairchem.core.datasets.atomic_data import AtomicData
 from fairchem.core.datasets.collaters.simple_collater import data_list_collater
-from fairchem.core.units.mlip_unit.mlip_unit import InferenceSettings, MLIPPredictUnit
+from fairchem.core.units.mlip_unit import MLIPPredictUnit
+from fairchem.core.units.mlip_unit.mlip_unit import InferenceSettings
 
 
 @pytest.mark.inference_check()
@@ -34,7 +35,8 @@ def test_inference_checkpoint_direct(
 
     db = AseDBDataset(config={"src": os.path.join(fake_uma_dataset, "oc20")})
 
-    a2g = partial(AtomicData.from_ase,
+    a2g = partial(
+        AtomicData.from_ase,
         max_neigh=10,
         radius=100,
         r_energy=False,
